+++ src/devtools/client/webconsole/actions/network.ts
@@ -1,20 +1,12 @@
 import { RequestInfo, RequestEventInfo } from "@recordreplay/protocol";
 import { ThreadFront } from "protocol/thread";
+import { NEW_NETWORK_REQUESTS } from "ui/actions/network";
+import { AppDispatch } from "ui/setup";
 
 export const setupNetwork = (store: any) => {
-  ThreadFront.findNetworkRequests(({ requests, events }) => {
-    store.dispatch(onNetworkRequestsThunk({ requests, events }));
-  });
+  ThreadFront.findNetworkRequests(data => store.dispatch(onNetworkRequestsThunk(data)));
 };
 
-const onNetworkRequestsThunk = ({
-  events,
-  requests,
-}: {
-  events: RequestEventInfo[];
-  requests: RequestInfo[];
-}) => {
-  return async ({ dispatch }: { dispatch: any }) => {
-    dispatch({ type: "NEW_NETWORK_REQUESTS", payload: { requests, events } });
-  };
+const onNetworkRequestsThunk = (data: { requests: RequestInfo[]; events: RequestEventInfo[] }) => {
+  async (dispatch: AppDispatch) => dispatch({ type: NEW_NETWORK_REQUESTS, payload: data });
 };
